refactor: split fastapi routers by domain
